Fixed a 5xx response when liking a post with an invalid ID

+++ SocialAppWebApi/Endpoints/LikesController.cs
@@ -8,13 +8,16 @@ namespace SocialAppWebApi.Endpoints;
 [Authorize]
 [ApiController]
 [Route("v1/[controller]")]
-public class LikesController(IUsersService usersService, ILikesService likesService) : UserControllerBase(usersService)
+public class LikesController(IUsersService usersService, ILikesService likesService, IPostsService postsService) : UserControllerBase(usersService)
 {
     [HttpPut]
     public async Task<IActionResult> CreateLike([FromBody] PostLikeDto likeDto)
     {
         if (await GetCurrentUserAsync() is not {} user)
             return Unauthorized();
+
+        if (await postsService.GetPostByIdAsync(likeDto.PostId) == null)
+            return NotFound();
         
         if (!await likesService.CreateLikeAsync(user, likeDto.PostId))
             return BadRequest("The current user has already liked this post");
@@ -27,6 +30,9 @@ public class LikesController(IUsersService usersService, ILikesService likesServ
     {
         if (await GetCurrentUserAsync() is not {} user)
             return Unauthorized();
+        
+        if (await postsService.GetPostByIdAsync(likeDto.PostId) == null)
+            return NotFound();
         
         if (!await likesService.DeleteLikeAsync(user, likeDto.PostId))
             return BadRequest("The current user has not liked this post yet");
